[refactor]페이지네이션 기능 공용 컴포넌트로 분리
