+++ src/utils/cookie.js
@@ -1,4 +1,4 @@
-import Cookies from "js-cookie";
+import Cookies from "js-cookie"; //서진님이 만든 쿠키~~너를 위해 구웠지~~
 
 /* 쿠키 저장 */
 export const setCookie = (key, value, options) => {
